Everything is a horrible mess, and this DOES NOT WORK.
Regardless, this is a good stopping point. I'm doing the subprocess thing,
with a Twisted reactor in another thread controlling a subprocess.

Given how things are, perhaps it's better to make the subprocess the server.
The only reason I don't want to do that is because it's very conceivable that
one might want multiple subprocesses later on (in particular, pygame).

Anyway, this is an incomplete mess, but it shows where I'm going.